requirements.txt
#==================================================================
# BMU API - Requirements
#==================================================================

# Async Web Framework
quart==0.20.0
quart-cors==0.8.0
hypercorn==0.18.0
motor==3.7.1

# HTTP & Web Scraping
httpx==0.28.1
beautifulsoup4==4.14.2

# Scheduling
apscheduler==3.11.1

# (Optional) Environment Variable Management
python-dotenv==1.2.1

# Data Validation
pydantic==2.10.3
